+++ src/app/test/page.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import { Box, Typography, Grid } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
-import NavigationBar from "@/feature/base/components/NavigationBar";
+import NavigationBar from "@/features/base/components/NavigationBar";
 
 export default function Home() {
   return <Box>Test</Box>;
